+++ src/com/jpexs/decompiler/flash/gui/generictageditors/StringEditor.java
@@ -23,6 +23,7 @@ import java.awt.Dimension;
 import java.awt.event.FocusAdapter;
 import java.awt.event.FocusEvent;
 import java.lang.reflect.Field;
+import java.util.Objects;
 import javax.swing.JTextArea;
 
 /**
@@ -91,12 +92,18 @@ public class StringEditor extends JTextArea implements GenericTagEditor {
     }
 
     @Override
-    public void save() {
+    public boolean save() {
         try {
+            String oldValue = (String) ReflectionTools.getValue(obj, field, index);
+            String newValue = getText();
+            if (Objects.equals(oldValue, newValue)) {
+                return false;
+            }
             ReflectionTools.setValue(obj, field, index, getText());
-        } catch (IllegalAccessException ex) {
+        } catch (IllegalArgumentException | IllegalAccessException ex) {
             // ignore
         }
+        return true;
     }
 
     @Override